+++ Packages/webxr-interactions/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Check if hand joints radius changed on hand update in ControllerInteraction.
 
 ## [0.5.2] - 2020-01-16
 ### Fixed